Make sure not to display an image that belongs to a different post
The key change here is to store the image subscription in the MediaView’s disposeBag rather than the configuration’s, so that it is properly disposed of in the MediaView’s prepareForReuse().

For good measure, also adding MainActor on prepareForReuse and actual cancellation of all subscriptions before releasing them.

Fixes #1374 [BUG] Posts can load the wrong images