Control the height of the compose window a little better.
Avoids the known hang, which was caused by getting stuck in a loop setting the frame back and forth over and over. Looked like it was adding and then removing (and then adding, etc) the height of a single line of text. Possible the username was shifting back and forth between truncating and wrapping, but have not confirmed that.

Fixes #705 IOS client consistently hangs when writing new post at ~= 150 chars