Avoid showing duplicate poll options
This is an invalid state, but currently it is possible to create a valid poll (no identical options) and later edit it so that two or more options contain identical strings. Bug (#33707) has been filed for the mastodon repository.

Fixes #1332 [BUG] Mastodon app crashes when a poll with identical answers reaches the screen