Resolve layout issue
Seems that having wrapped text in the header line could cause the post preview to show fewer than the allowed 3 lines, but there was still space provided for the lines expected. This caused the layout of the whole row to get messed up, with the avatar row pushed down, out of horizontal alignment with the notification icon.

Contributes to #399 [BUG] Multiple interactions do not collapse into a single notification